@@ -11,45 +11,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.0.8",
-    "@angular/common": "^10.0.8",
-    "@angular/compiler": "^10.0.8",
-    "@angular/core": "^10.0.8",
-    "@angular/forms": "^10.0.8",
-    "@angular/platform-browser": "^10.0.8",
-    "@angular/platform-browser-dynamic": "^10.0.8",
-    "@angular/router": "^10.0.8",
-    "@ngrx/data": "^10.0.0",
-    "@ngrx/effects": "^10.0.0",
-    "@ngrx/entity": "^10.0.0",
-    "@ngrx/router-store": "^10.0.0",
-    "@ngrx/store": "^10.0.0",
-    "@ngrx/store-devtools": "^10.0.0",
-    "ngrx-undo": "^8.0.0",
-    "rxjs": "^6.6.2",
-    "tslib": "^2.0.1",
-    "zone.js": "^0.10.3"
+    "@angular/animations": "~11.2.0",
+    "@angular/cdk": "^11.2.0",
+    "@angular/common": "~11.2.0",
+    "@angular/compiler": "~11.2.0",
+    "@angular/core": "~11.2.0",
+    "@angular/flex-layout": "^11.0.0-beta.33",
+    "@angular/forms": "~11.2.0",
+    "@angular/material": "^11.2.0",
+    "@angular/platform-browser": "~11.2.0",
+    "@angular/platform-browser-dynamic": "~11.2.0",
+    "@angular/router": "~11.2.0",
+    "@ngrx/component-store": "^11.0.0",
+    "@ngrx/data": "^11.0.0",
+    "@ngrx/effects": "^11.0.0",
+    "@ngrx/entity": "^11.0.0",
+    "@ngrx/router-store": "^11.0.0",
+    "@ngrx/store": "^11.0.0",
+    "@ngrx/store-devtools": "^11.0.0",
+    "ngrx-forms": "^6.3.4",
+    "rxjs": "~6.6.0",
+    "tslib": "^2.0.0",
+    "zone.js": "~0.11.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1000.5",
-    "@angular/cli": "^10.0.5",
-    "@angular/compiler-cli": "^10.0.8",
-    "@angular/language-service": "^10.0.8",
-    "@types/jasmine": "^3.5.12",
+    "@angular-devkit/build-angular": "~0.1102.0",
+    "@angular/cli": "~11.2.0",
+    "@angular/compiler-cli": "~11.2.0",
+    "@angular/language-service": "^11.2.0",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "^2.0.8",
-    "@types/node": "^14.0.27",
+    "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
-    "eslint": "^7.6.0",
     "jasmine-core": "^3.6.0",
-    "jasmine-spec-reporter": "^5.0.2",
-    "karma": "^5.1.1",
-    "karma-chrome-launcher": "^3.1.0",
-    "karma-coverage-istanbul-reporter": "^3.0.3",
-    "karma-jasmine": "^3.3.1",
-    "karma-jasmine-html-reporter": "^1.5.4",
+    "jasmine-spec-reporter": "^5.0.0",
+    "karma": "~6.1.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.0.3",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "protractor": "^7.0.0",
-    "ts-node": "^8.10.2",
-    "tslint": "^6.1.3",
-    "typescript": "~3.9.7"
+    "ts-node": "~8.3.0",
+    "tslint": "~6.1.0",
+    "typescript": "~4.1.2"
   }
 }
