@@ -6,17 +6,20 @@ import {HeroResolver} from '../resolvers/hero.resolver';
 import {NgModule} from '@angular/core';
 
 const heroesRoutes: Routes = [
-    { path: '', children: [
-        {path: '', redirectTo: 'heroes', pathMatch: 'full'},
-        {path: 'heroes', component: HeroesListComponent, resolve: {entity: HeroesResolver}},
-        {path: ':id', component: HeroDetailComponent, resolve: {entity: HeroResolver}},
-    ]},
+  {
+    path: '', children: [
+      {path: '', redirectTo: 'heroes', pathMatch: 'full'},
+      {path: 'heroes', component: HeroesListComponent, resolve: {entity: HeroesResolver}},
+      {path: ':id', component: HeroDetailComponent, resolve: {entity: HeroResolver}},
+    ]
+  },
 ];
 
 @NgModule({
-    imports: [
-        RouterModule.forChild(heroesRoutes)
-    ],
-    exports: [RouterModule]
- })
-export class HeroesRoutingModule {}
+  imports: [
+    RouterModule.forChild(heroesRoutes)
+  ],
+  exports: [RouterModule]
+})
+export class HeroesRoutingModule {
+}
