@@ -1,4 +1,10 @@
 export class Hero {
-    id?: string;
-    name: string;
+  id?: string;
+  name: string;
+  // tslint:disable-next-line:variable-name
+  full_name: string;
+  thumb: string;
+  photo: string;
+  description: string;
+  text: string;
 }
