@@ -6,22 +6,24 @@ import {StoreModule} from '@ngrx/store';
 import {routerReducer, StoreRouterConnectingModule} from '@ngrx/router-store';
 import {StoreDevtoolsModule} from '@ngrx/store-devtools';
 import {environment} from '../../environments/environment';
+import {HeroFormStoreModule} from './hero-form-store';
 
 @NgModule({
-    declarations: [],
-    imports: [
-        CommonModule,
-        HeroStoreModule,
-        EffectsModule.forRoot([]),
-        StoreModule.forRoot({
-            router: routerReducer,
-        }),
-        StoreRouterConnectingModule.forRoot(),
-        StoreDevtoolsModule.instrument({
-            maxAge: 25,
-            logOnly: environment.production
-        }),
-    ]
+  declarations: [],
+  imports: [
+    CommonModule,
+    HeroStoreModule,
+    HeroFormStoreModule,
+    EffectsModule.forRoot([]),
+    StoreModule.forRoot({
+      router: routerReducer,
+    }),
+    StoreRouterConnectingModule.forRoot(),
+    StoreDevtoolsModule.instrument({
+      maxAge: 25,
+      logOnly: environment.production
+    }),
+  ]
 })
 export class RootStoreModule {
 }
